JP-478: one home for byte formatting
`formatFileSize` existed twice with different behaviour — the `imageUtils` copy
stopped at MB (a 20 GB quota rendered as `20480.0 MB`), the `fileUtils` copy
continued to GB and guarded bad input. Which one a screen got was decided by an
import, so the same byte count could read differently on two surfaces of the
same app.

Both are gone. `src/utils/byteSize.ts` (renamed from JP-477's
`formatStorageSize.ts`) is now the single home, exporting one ladder under two
presentations:

- `formatFileSize` — how big is this *thing*. One decimal, because `1.50 GB` is
  noise on a file.
- `formatStorageSize` — how much of an *allowance* is left. Two decimals from GB
  up, so a filling workspace moves the number instead of sitting at `7.7 GB`.

They stay two functions because they answer different questions, but they share
the ladder, so a unit boundary can only be wrong in one place. That is the
argument for merging the modules rather than leaving `formatFileSize` in
`fileUtils` and having it call across — the duplication worth removing was the
ladder, not the entry point.

**Rounding fix that falls out of sharing it.** Both old implementations printed
`1024.0 KB` for one byte short of a megabyte: the value rounds up onto the very
boundary the rung was chosen to be below. The shared ladder promotes instead, so
that reads `1.0 MB`. Narrow — a one-byte window per boundary — but it was in
both copies, which is the point.

Eleven call sites re-pointed. Two of them (`FileViewerContent`,
`GenericFileViewer`) split their import, since they were pulling
`formatFileSize` alongside genuine `fileUtils` exports.

Visible change: the six surfaces that used the MB-capped copy now render
gigabyte-scale values as GB. Confirmed live — Storage Management's "Available"
reads `132.1 GB` where it read `135266.5 MB`. The image pickers and PDF dialog
never see a gigabyte, so nothing moves there.

Verified: typecheck, 3356 tests across 285 files, production build; and against
the running local stack for the two surfaces whose output actually changes.
