+++ src/utils/fileUtils.test.ts
@@ -1,7 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import {
   detectFileCategory,
-  formatFileSize,
   getMimeType,
   isPreviewableFile,
   resolveViewerCategory,
@@ -68,31 +67,6 @@ describe('resolveViewerCategory', () => {
 
   it('keeps truly generic files generic', () => {
     expect(resolveViewerCategory('generic', 'application/octet-stream')).toBe('generic');
-  });
-});
-
-describe('formatFileSize', () => {
-  it('formats bytes', () => {
-    expect(formatFileSize(0)).toBe('0 B');
-    expect(formatFileSize(512)).toBe('512 B');
-  });
-
-  it('formats kilobytes', () => {
-    expect(formatFileSize(1024)).toBe('1.0 KB');
-    expect(formatFileSize(1536)).toBe('1.5 KB');
-  });
-
-  it('formats megabytes', () => {
-    expect(formatFileSize(1024 * 1024)).toBe('1.0 MB');
-    expect(formatFileSize(5.5 * 1024 * 1024)).toBe('5.5 MB');
-  });
-
-  it('formats gigabytes', () => {
-    expect(formatFileSize(1024 * 1024 * 1024)).toBe('1.0 GB');
-  });
-
-  it('handles negative values', () => {
-    expect(formatFileSize(-1)).toBe('0 B');
   });
 });
