+++ src/utils/fileUtils.ts
@@ -93,18 +93,6 @@ export function resolveViewerCategory(category: FileCategory, mimeType: string):
 }
 
 /**
- * Format a file size in bytes to a human-readable string.
- */
-export function formatFileSize(bytes: number): string {
-  if (bytes < 0) return '0 B';
-  if (bytes === 0) return '0 B';
-  if (bytes < 1024) return `${bytes} B`;
-  if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
-  if (bytes < 1024 * 1024 * 1024) return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
-  return `${(bytes / (1024 * 1024 * 1024)).toFixed(1)} GB`;
-}
-
-/**
  * Guess MIME type from filename extension.
  * Returns 'application/octet-stream' for unknown types.
  */
