+++ src/utils/imageUtils.ts
@@ -165,18 +165,6 @@ export async function processImageForUpload(file: File): Promise<{
 }
 
 /**
- * Format file size for display.
- *
- * @param bytes - File size in bytes
- * @returns Formatted string (e.g., "1.5 MB")
- */
-export function formatFileSize(bytes: number): string {
-  if (bytes < 1024) return `${bytes} B`;
-  if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
-  return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
-}
-
-/**
  * Create a thumbnail from an image blob.
  * Useful for preview UI.
  *
